+++ Controllers/CASCController.cs
@@ -1036,30 +1036,38 @@ namespace wow.tools.local.Controllers
 
             if (CASC.FDIDToCHash.TryGetValue(filedataid, out var cKeyBytes))
             {
-                if (CASC.FDIDToExtraCHashes.TryGetValue(filedataid, out List<byte[]>? extraCHashes))
+                var primaryCKey = Convert.ToHexStringLower(cKeyBytes);
+
+                var allCKeys = CASC.GetCKeysAndFlagsByFDID(filedataid);
+                html += "<tr><td>Content Hashes</td><td>";
+                html += "<table class='table table-sm table-inverse'>";
+                html += "<thead><tr><th>MD5/CKey</th><th>LocaleFlags</th><th>ContentFlags</th><th>&nbsp;</th></tr></thead>";
+
+                foreach (var extraCKeyEntry in allCKeys)
                 {
-                    var cKey = Convert.ToHexStringLower(cKeyBytes);
-                    html += "<tr><td>Content hash (MD5)</td><td style='font-family: monospace;'>";
-
-                    html += "<a href='#' data-bs-toggle='modal' data-bs-target='#chashModal' onClick='fillChashModal(\"" + cKey.ToLower() + "\")'>" + cKey.ToLower() + "</a> (preferred)<br>";
-
-                    foreach (var extraCKey in extraCHashes)
-                    {
-                        var extraCKeyHex = Convert.ToHexStringLower(extraCKey);
-                        html += "<a href='#' data-bs-toggle='modal' data-bs-target='#chashModal' onClick='fillChashModal(\"" + extraCKeyHex.ToLower() + "\")'>" + extraCKeyHex.ToLower() + "</a> (<a href='/casc/chash?contenthash=" + extraCKeyHex + "&filename=" + filedataid + ".bytes'>download</a>)<br>";
-                    }
-
-                    html += "</td></tr>";
-
-                    html += "<tr><td>Size</td><td>" + (CASC.CHashToSize.TryGetValue(cKey, out var size) ? size + " bytes" : "N/A") + " (for preferred version)</td></tr>";
-                }
-                else
-                {
-                    var cKey = Convert.ToHexStringLower(cKeyBytes);
-                    html += "<tr><td>Content hash (MD5)</td><td style='font-family: monospace;'><a href='#' data-bs-toggle='modal' data-bs-target='#chashModal' onClick='fillChashModal(\"" + cKey.ToLower() + "\")'>" + cKey.ToLower() + "</a></td></tr>";
-                    html += "<tr><td>Size</td><td>" + (CASC.CHashToSize.TryGetValue(cKey, out var size) ? size + " bytes" : "N/A") + "</td></tr>";
+                    var extraCKeyHex = Convert.ToHexStringLower(extraCKeyEntry.cKey);
+                 
+                    html += "<tr>";
+                    html += "<td style='font-family: monospace;'><a href='#' data-bs-toggle='modal' data-bs-target='#chashModal' onClick='fillChashModal(\"" + extraCKeyHex + "\")'>" + extraCKeyHex + "</a>";
+                    if(primaryCKey == extraCKeyHex && allCKeys.Count > 1)
+                        html += " (preferred)";
+                    html += "</td>"; 
+                    html += "<td>" + extraCKeyEntry.localeFlags + "</td>";
+                    html += "<td>" + extraCKeyEntry.contentFlags + "</td>";
+                    html += "<td><a href='/casc/chash?contenthash=" + extraCKeyHex + "&filename=" + filedataid + ".bytes'>download</a></td>";
+                    html += "</tr>";
                 }
 
+                html += "</table>";
+
+                html += "</td></tr>";
+
+                html += "<tr><td>Size</td><td>" + (CASC.CHashToSize.TryGetValue(primaryCKey, out var size) ? size + " bytes" : "N/A");
+
+                if (allCKeys.Count > 1)
+                    html += " (for preferred version)";
+                
+                html += "</td></tr>";
             }
 
             if (CASC.EncryptionStatuses.TryGetValue(filedataid, out var encryptionStatus))